Refactored utility methods for DL api to lib file
